# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Java Platform API
  version: 1.0.0
  description: |
    Task management API secured with JWT bearer tokens.

    Obtain a token from `POST /api/v1/auth/login`, then send it as
    `Authorization: Bearer <token>` on every `/api/v1/tasks` request.

    This document is committed for review and client generation. The **runtime**
    document is generated by springdoc from the annotated controllers and served
    at `/v3/api-docs` — that one is authoritative if the two ever disagree.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://{host}
    description: Deployed instance (nginx on port 80)
    variables:
      host:
        default: elastic-ip
        description: The instance Elastic IP or DNS name

tags:
  - name: Authentication
    description: JWT token issuance
  - name: Tasks
    description: Task management operations
  - name: Actuator
    description: Health and build information

paths:
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Exchange credentials for a JWT bearer token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid credentials (no body)

  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: List tasks, optionally filtered by status
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskStatus'
          description: Case-insensitive. Omit for all tasks.
      responses:
        '200':
          description: Matching tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: URI of the created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/tasks/urgent:
    get:
      tags: [Tasks]
      summary: List tasks at priority 1 or 2
      responses:
        '200':
          description: Urgent tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Tasks]
      summary: Fetch a single task
      responses:
        '200':
          description: The task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Tasks]
      summary: Replace a task's editable fields
      description: Title, description and priority. Status is changed via the status endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Tasks]
      summary: Delete a task
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/tasks/{id}/status:
    patch:
      tags: [Tasks]
      summary: Transition a task to a new status
      description: |
        A task in `DONE` is terminal: any transition away from it is rejected
        with `409` so completed work cannot be silently reopened by a retrying
        client. Re-applying `DONE` to a `DONE` task is idempotent.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: value
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/TaskStatus'
      responses:
        '200':
          description: Transitioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Illegal transition (a completed task cannot be reopened)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemResponse'

  /actuator/health:
    get:
      tags: [Actuator]
      summary: Health probe including PostgreSQL connectivity
      security: []
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: UP
                  components:
                    type: object
        '503':
          description: Unhealthy — inspect the failing component

  /actuator/info:
    get:
      tags: [Actuator]
      summary: Build, JVM and OS information
      security: []
      responses:
        '200':
          description: Application information
          content:
            application/json:
              schema:
                type: object

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    TaskStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, BLOCKED, DONE]

    AuthRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: operator
        password:
          type: string
          format: password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: HS256 JWT
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          description: Lifetime in seconds
          example: 3600

    TaskRequest:
      type: object
      required: [title, priority]
      properties:
        title:
          type: string
          maxLength: 200
          example: Ship the release pipeline
        description:
          type: string
          maxLength: 2000
          nullable: true
        priority:
          type: integer
          minimum: 1
          maximum: 5
          example: 2

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProblemResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
          description: Present only on validation failures

  responses:
    Unauthorized:
      description: Missing, malformed, expired or forged token (no body)
    NotFound:
      description: No task with that id
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemResponse'
    BadRequest:
      description: Malformed request, e.g. an unknown status value
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemResponse'
    ValidationError:
      description: Bean validation failed; `fields` names each rejected field
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemResponse'

security:
  - bearerAuth: []